requirements.txt
discord==1.7.3
python-dotenv==0.19.2
ujson==5.1.0
